Faz 1 adim 6a: admin backend — uye yonetimi + agac + dashboard
Admin web'in API bagimliliklari (SPEC 9):
- /admin/members: GET list (arama/durum/sayfali), GET tree (parent/depth gorsel),
  POST invite (sponsor secili, varsayilan admin), POST :id/deactivate|activate,
  POST :id/role (tenant_admin|staff|member; owner degistirilemez) — audit'li
- /admin/dashboard: ay ciro, komisyon gideri, efektif oran (bps), uye sayilari,
  acik payable bakiye, bekleyen payout talebi
- /admin/audit: tenant audit log (admin+)
- RBAC: list/tree/dashboard staff+; davet/rol/durum/audit admin+; tenant-scope (404)

Duzeltme: raw SUM(bigint) Postgres'te numeric doner; ::bigint cast eklendi
(reports dashboard + payouts payable) — BigInt karistirma hatasi giderildi.

Test: 48 entegrasyon (6 yeni admin senaryosu) yesil.

+++ apps/api/src/payouts/payouts.service.ts
@@ -27,7 +27,7 @@ export class PayoutsService {
       SELECT le.beneficiary_membership_id AS "membershipId",
              m.referral_code              AS "referralCode",
              u.full_name                  AS "fullName",
-             SUM(le.amount_cents)         AS "netCents"
+             SUM(le.amount_cents)::bigint AS "netCents"
       FROM ledger_entries le
       JOIN memberships m ON m.id = le.beneficiary_membership_id
       JOIN users u       ON u.id = m.user_id
